+++ Sources/NFCDemo/NFC/NFCLoginController.swift
@@ -133,7 +133,13 @@ public class NFCLoginController: LoginController {
             .receive(on: DispatchQueue.main)
             .sink(receiveCompletion: { [weak self] completion in
                 if case let .failure(error) = completion {
-                    self?.pState = .error(error)
+                    if case let .cardError(readerError) = error as? NFCLoginController.Error,
+                       case let .nfcTag(error: tagError) = readerError,
+                       case .userCanceled = tagError {
+                        self?.pState = .idle
+                    } else {
+                        self?.pState = .error(error)
+                    }
                 } else {
                     self?.pState = .idle
                 }