+++ Sources/NFCDemo/NFC/NFCResetRetryCounterController.swift
@@ -132,7 +132,13 @@ public class NFCResetRetryCounterController: ResetRetryCounter {
             .sink(
                 receiveCompletion: { [weak self] completion in
                     if case let .failure(error) = completion {
-                        self?.pState = .error(error)
+                        if case let .cardError(readerError) = error as? NFCResetRetryCounterController.Error,
+                           case let .nfcTag(error: tagError) = readerError,
+                           case .userCanceled = tagError {
+                            self?.pState = .idle
+                        } else {
+                            self?.pState = .error(error)
+                        }
                     } else {
                         self?.pState = .idle
                     }