+++ Tests/NFCDemoTests/RegisterCANViewSnapshotTests.swift
@@ -20,6 +20,12 @@ import SwiftUI
 import XCTest
 
 class RegisterCANViewSnapshotTests: XCTestCase {
+    override func setUp() {
+        super.setUp()
+
+        diffTool = "open"
+    }
+
     func testRegisterCANViewSnapshotTests() throws {
         let sut = RegisterCANView()
             .frame(width: UIScreen.main.bounds.width, height: UIScreen.main.bounds.height)